chore(boundary): total agent/supervisor import separation
Extract the benchmark / --run-test-instance / --run-fake-instance
harnesses out of aleph.vm.agent.cli into aleph.vm.testing.harness. They
wire BOTH sides in one process (VmPool + LocalSupervisor + the agent's
run/registry/expiry machinery), so they cannot live inside the agent
package; the cli loads them lazily, only when one of those flags is
given, and the production agent path never touches supervisor-side code.

With the wire client on the contract side and the harness out of the
agent, the import-linter contracts become total, with zero ignores:

- aleph.vm.agent imports neither aleph.vm.supervisor nor aleph.vm.pool
  nor aleph.vm.network (network is supervisor-side host plumbing)
- the supervisor side (supervisor, pool, models, network, hypervisors)
  does not import the agent or the harness; the per-submodule
  'controllers' contract is subsumed and dropped
- aleph.vm.supervisor_interface still imports neither side

Remaining shared-domain edges, documented in the contract and left for
the model split: agent -> models (VmExecution, MigrationState), agent ->
hypervisors (two leaf classes), agent -> migration (agent-side jobs).
